This commit was generated by cvs2svn to compensate for changes in r7753, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@7754 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/uimon.h
@@ -46,12 +46,14 @@ extern char *            uimon_get_in         (char **);
 
 #ifdef WIN32
 
+#include "types.h"
+
 /*
  For the time being, this internal interface remains here...
 */
 extern struct console_s *uimon_console_open_mdi(const char *id, void *,
                                                 void *, void *,
-                                                unsigned long int dwStyle,
+                                                DWORD dwStyle,
                                                 int x, int y, int dx, int dy );
 #endif /* #ifdef WIN32 */
